requirements_.txt
albumentations==1.4.1
packaging==23.2
Pillow==10.0.0
PyYAML==6.0.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
torch==2.0.1+cu117
torchsummary==1.5.1
torchvision==0.15.2+cu117
tornado==6.4
tqdm==4.66.1
